Aggregate an imputed dataset

Aggregate an imputed dataset

aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for ANY
aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for rawImputed
aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for aggregatedImputed
aggregate_impute(object, grouping, fun, filter = list(), join)

Arguments

object

A rawImputed object.

grouping

A vector of variables names to group the aggregation on.

fun

The function to aggregate.

filter

An optional argument to filter the raw dataset before aggregation. Will be passed to dplyr::filter().

join

An optional argument to filter the raw dataset based on a data.frame. A dplyr::semi_join() will be applied with join or each element of join in case join is a list.

Examples

dataset <- generate_data(n_year = 10, n_site = 50, n_run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
#> An object of class "aggregatedImputed"
#> Slot "Covariate":
#>    Year Period
#> 1     1      1
#> 2     1      2
#> 3     1      3
#> 4     1      4
#> 5     1      5
#> 6     1      6
#> 7     2      1
#> 8     2      2
#> 9     2      3
#> 10    2      4
#> 11    2      5
#> 12    2      6
#> 13    3      1
#> 14    3      2
#> 15    3      3
#> 16    3      4
#> 17    3      5
#> 18    3      6
#> 19    4      1
#> 20    4      2
#> 21    4      3
#> 22    4      4
#> 23    4      5
#> 24    4      6
#> 25    5      1
#> 26    5      2
#> 27    5      3
#> 28    5      4
#> 29    5      5
#> 30    5      6
#> 31    6      1
#> 32    6      2
#> 33    6      3
#> 34    6      4
#> 35    6      5
#> 36    6      6
#> 37    7      1
#> 38    7      2
#> 39    7      3
#> 40    7      4
#> 41    7      5
#> 42    7      6
#> 43    8      1
#> 44    8      2
#> 45    8      3
#> 46    8      4
#> 47    8      5
#> 48    8      6
#> 49    9      1
#> 50    9      2
#> 51    9      3
#> 52    9      4
#> 53    9      5
#> 54    9      6
#> 55   10      1
#> 56   10      2
#> 57   10      3
#> 58   10      4
#> 59   10      5
#> 60   10      6
#> 
#> Slot "Imputation":
#>       Imputation0001 Imputation0002 Imputation0003 Imputation0004
#>  [1,]      1087.8288      1086.9465      1134.2921      1117.2576
#>  [2,]      1624.0000      1624.0000      1624.0000      1624.0000
#>  [3,]      1913.0000      1913.0000      1913.0000      1913.0000
#>  [4,]      1402.2731      1496.9259      1441.3542      1412.7013
#>  [5,]       886.0000       886.0000       886.0000       886.0000
#>  [6,]       651.9220       609.3896       633.0872       569.5151
#>  [7,]      1034.0000      1034.0000      1034.0000      1034.0000
#>  [8,]      1650.4517      1617.4266      1615.0208      1584.6080
#>  [9,]      1630.0000      1630.0000      1630.0000      1630.0000
#> [10,]      1103.0000      1103.0000      1103.0000      1103.0000
#> [11,]      1168.0000      1168.0000      1168.0000      1168.0000
#> [12,]       403.0000       403.0000       403.0000       403.0000
#> [13,]      1154.2099      1112.1392      1157.1854      1101.7413
#> [14,]      1573.9568      1511.4400      1622.0861      1584.5161
#> [15,]      1390.0781      1397.4096      1379.4488      1364.5321
#> [16,]      1243.9584      1231.0714      1264.8551      1260.3855
#> [17,]      1045.1062      1057.2773       994.4441       988.9188
#> [18,]       331.6656       417.1328       385.4496       450.9484
#> [19,]      1320.9275      1460.9350      1328.8972      1483.1940
#> [20,]      1835.0000      1835.0000      1835.0000      1835.0000
#> [21,]      1584.0000      1584.0000      1584.0000      1584.0000
#> [22,]      1592.9554      1579.7759      1624.8935      1531.1724
#> [23,]       834.6112       953.4034       810.7872       917.9240
#> [24,]       522.2141       514.5517       484.9197       610.8934
#> [25,]      1102.0000      1102.0000      1102.0000      1102.0000
#> [26,]      1669.0000      1669.0000      1669.0000      1669.0000
#> [27,]      1528.2606      1535.4693      1601.5864      1516.5961
#> [28,]      1214.5480      1125.1489      1143.1577      1117.1176
#> [29,]       573.5535       612.6434       591.4485       594.0853
#> [30,]       538.1435       544.1808       541.1121       440.6750
#> [31,]      1515.0478      1548.8847      1596.7301      1473.3266
#> [32,]      1712.2318      1719.7631      1589.9898      1597.5209
#> [33,]      1389.0000      1389.0000      1389.0000      1389.0000
#> [34,]      1144.0000      1144.0000      1144.0000      1144.0000
#> [35,]       742.2116       785.8371       740.0428       735.4509
#> [36,]       548.1429       470.3017       562.8128       444.1222
#> [37,]       876.0000       876.0000       876.0000       876.0000
#> [38,]      1827.0000      1827.0000      1827.0000      1827.0000
#> [39,]      1804.7797      1719.4075      1783.4089      1708.7849
#> [40,]      1750.2754      1803.2295      1698.0796      1726.2491
#> [41,]      1539.0000      1539.0000      1539.0000      1539.0000
#> [42,]       982.0000       982.0000       982.0000       982.0000
#> [43,]       770.2689       764.6447       766.3184       724.9371
#> [44,]      1860.9389      1917.0675      1940.3343      1969.9327
#> [45,]      2218.0000      2218.0000      2218.0000      2218.0000
#> [46,]      1451.5939      1512.1440      1452.6242      1471.5991
#> [47,]      1577.0000      1577.0000      1577.0000      1577.0000
#> [48,]       813.0000       813.0000       813.0000       813.0000
#> [49,]       661.8807       754.0529       771.1503       733.8477
#> [50,]      1565.7648      1477.6892      1519.5643      1462.0026
#> [51,]      1566.6618      1617.3315      1628.4244      1654.8209
#> [52,]      1489.0000      1489.0000      1489.0000      1489.0000
#> [53,]      1088.0000      1088.0000      1088.0000      1088.0000
#> [54,]       745.0000       745.0000       745.0000       745.0000
#> [55,]      1052.3837      1021.9246      1185.1376      1250.8264
#> [56,]      1628.4062      1608.1835      1604.2780      1586.2062
#> [57,]      1543.0866      1519.3465      1572.0315      1575.6163
#> [58,]      1693.0000      1693.0000      1693.0000      1693.0000
#> [59,]      1551.5237      1674.6076      1545.4662      1493.6745
#> [60,]       728.8550       692.8864       828.5285       911.5440
#>       Imputation0005 Imputation0006 Imputation0007 Imputation0008
#>  [1,]      1118.3009      1124.9967      1053.2544      1091.6772
#>  [2,]      1624.0000      1624.0000      1624.0000      1624.0000
#>  [3,]      1913.0000      1913.0000      1913.0000      1913.0000
#>  [4,]      1410.1294      1410.6914      1432.3120      1403.6249
#>  [5,]       886.0000       886.0000       886.0000       886.0000
#>  [6,]       695.3980       632.2228       673.0899       596.7047
#>  [7,]      1034.0000      1034.0000      1034.0000      1034.0000
#>  [8,]      1575.7640      1596.0603      1587.8059      1720.0528
#>  [9,]      1630.0000      1630.0000      1630.0000      1630.0000
#> [10,]      1103.0000      1103.0000      1103.0000      1103.0000
#> [11,]      1168.0000      1168.0000      1168.0000      1168.0000
#> [12,]       403.0000       403.0000       403.0000       403.0000
#> [13,]      1095.4562      1117.4184      1058.8527      1147.1060
#> [14,]      1496.0309      1433.3892      1470.6432      1535.1837
#> [15,]      1463.3851      1383.3539      1388.2057      1448.4003
#> [16,]      1374.7652      1226.8133      1188.0561      1221.9101
#> [17,]       956.0330      1036.5651      1036.0944       874.3921
#> [18,]       446.9482       424.8778       411.8089       478.9685
#> [19,]      1443.8706      1425.2594      1244.7198      1439.0140
#> [20,]      1835.0000      1835.0000      1835.0000      1835.0000
#> [21,]      1584.0000      1584.0000      1584.0000      1584.0000
#> [22,]      1536.1086      1578.6390      1546.2782      1581.8217
#> [23,]       938.1862       951.0052       898.9720       878.5845
#> [24,]       508.9896       549.3627       538.2816       551.2404
#> [25,]      1102.0000      1102.0000      1102.0000      1102.0000
#> [26,]      1669.0000      1669.0000      1669.0000      1669.0000
#> [27,]      1579.4526      1597.4128      1513.1844      1585.3204
#> [28,]      1134.0498      1170.0766      1218.6958      1151.2504
#> [29,]       719.8886       638.0324       650.3559       532.9711
#> [30,]       555.1761       630.3775       572.2073       489.5667
#> [31,]      1566.5627      1545.6993      1500.8071      1548.6275
#> [32,]      1637.7715      1598.4880      1685.3503      1655.1310
#> [33,]      1389.0000      1389.0000      1389.0000      1389.0000
#> [34,]      1144.0000      1144.0000      1144.0000      1144.0000
#> [35,]       781.6760       783.0402       717.4017       685.2185
#> [36,]       564.0443       520.9379       513.6286       521.5319
#> [37,]       876.0000       876.0000       876.0000       876.0000
#> [38,]      1827.0000      1827.0000      1827.0000      1827.0000
#> [39,]      1706.9047      1739.7179      1719.2142      1778.8501
#> [40,]      1712.6380      1759.7094      1775.1325      1754.3794
#> [41,]      1539.0000      1539.0000      1539.0000      1539.0000
#> [42,]       982.0000       982.0000       982.0000       982.0000
#> [43,]       783.1890       721.4219       706.5257       780.5449
#> [44,]      1946.6335      1959.5119      1845.9648      1926.1450
#> [45,]      2218.0000      2218.0000      2218.0000      2218.0000
#> [46,]      1465.9603      1495.8713      1487.7785      1539.9734
#> [47,]      1577.0000      1577.0000      1577.0000      1577.0000
#> [48,]       813.0000       813.0000       813.0000       813.0000
#> [49,]       740.1449       715.3686       696.4716       769.5276
#> [50,]      1503.5568      1520.8451      1494.6551      1535.6231
#> [51,]      1648.9630      1615.7218      1655.0459      1687.0903
#> [52,]      1489.0000      1489.0000      1489.0000      1489.0000
#> [53,]      1088.0000      1088.0000      1088.0000      1088.0000
#> [54,]       745.0000       745.0000       745.0000       745.0000
#> [55,]      1094.6344      1112.1218      1149.0603      1075.3822
#> [56,]      1632.6952      1655.5665      1598.7993      1652.9478
#> [57,]      1601.9550      1591.8328      1586.6527      1578.4708
#> [58,]      1693.0000      1693.0000      1693.0000      1693.0000
#> [59,]      1575.2974      1580.1961      1584.3410      1570.4967
#> [60,]       815.0286       824.4863       847.9475       833.6832
#>       Imputation0009 Imputation0010 Imputation0011 Imputation0012
#>  [1,]      1046.8856      1160.6489      1027.0912      1160.1754
#>  [2,]      1624.0000      1624.0000      1624.0000      1624.0000
#>  [3,]      1913.0000      1913.0000      1913.0000      1913.0000
#>  [4,]      1509.0624      1454.9628      1518.0950      1481.1167
#>  [5,]       886.0000       886.0000       886.0000       886.0000
#>  [6,]       632.7903       687.2370       700.2878       618.4476
#>  [7,]      1034.0000      1034.0000      1034.0000      1034.0000
#>  [8,]      1613.6764      1583.8233      1667.7664      1636.2529
#>  [9,]      1630.0000      1630.0000      1630.0000      1630.0000
#> [10,]      1103.0000      1103.0000      1103.0000      1103.0000
#> [11,]      1168.0000      1168.0000      1168.0000      1168.0000
#> [12,]       403.0000       403.0000       403.0000       403.0000
#> [13,]      1093.4381      1160.5818      1051.1184      1117.7079
#> [14,]      1517.3275      1543.6254      1605.5033      1612.0845
#> [15,]      1389.7507      1421.1491      1415.7895      1376.3854
#> [16,]      1217.3025      1251.1579      1228.3136      1186.0925
#> [17,]       976.2913       904.8851       960.0211       929.8437
#> [18,]       434.3057       415.6852       469.2620       407.3705
#> [19,]      1473.9710      1499.8482      1445.6167      1370.4669
#> [20,]      1835.0000      1835.0000      1835.0000      1835.0000
#> [21,]      1584.0000      1584.0000      1584.0000      1584.0000
#> [22,]      1576.3121      1583.1618      1610.3034      1601.3201
#> [23,]       919.2901       904.5819       870.3557       934.4032
#> [24,]       512.0724       553.3383       504.0694       458.1675
#> [25,]      1102.0000      1102.0000      1102.0000      1102.0000
#> [26,]      1669.0000      1669.0000      1669.0000      1669.0000
#> [27,]      1545.2800      1532.4979      1532.5416      1501.9275
#> [28,]      1170.1628      1146.2597      1199.4936      1158.6762
#> [29,]       605.3868       565.8875       660.9455       648.7721
#> [30,]       562.3555       580.0542       484.8815       530.0347
#> [31,]      1569.0429      1497.1895      1556.3107      1569.0786
#> [32,]      1698.8890      1686.7086      1639.4232      1631.2065
#> [33,]      1389.0000      1389.0000      1389.0000      1389.0000
#> [34,]      1144.0000      1144.0000      1144.0000      1144.0000
#> [35,]       755.4314       669.9483       762.7332       719.1488
#> [36,]       553.8959       542.9005       625.8418       548.1820
#> [37,]       876.0000       876.0000       876.0000       876.0000
#> [38,]      1827.0000      1827.0000      1827.0000      1827.0000
#> [39,]      1729.5040      1698.6765      1761.6274      1732.8531
#> [40,]      1810.1657      1685.6670      1709.1756      1709.8287
#> [41,]      1539.0000      1539.0000      1539.0000      1539.0000
#> [42,]       982.0000       982.0000       982.0000       982.0000
#> [43,]       708.0581       723.7662       762.0104       722.7789
#> [44,]      1985.1151      1911.0264      1928.2833      1914.1195
#> [45,]      2218.0000      2218.0000      2218.0000      2218.0000
#> [46,]      1508.7638      1511.2291      1471.7644      1531.4800
#> [47,]      1577.0000      1577.0000      1577.0000      1577.0000
#> [48,]       813.0000       813.0000       813.0000       813.0000
#> [49,]       651.9895       790.2520       684.4348       703.3731
#> [50,]      1516.7491      1539.4557      1544.1087      1522.0231
#> [51,]      1599.2932      1599.2423      1579.9207      1619.8081
#> [52,]      1489.0000      1489.0000      1489.0000      1489.0000
#> [53,]      1088.0000      1088.0000      1088.0000      1088.0000
#> [54,]       745.0000       745.0000       745.0000       745.0000
#> [55,]      1117.7836      1074.4990      1230.8861      1126.3181
#> [56,]      1623.4935      1661.4193      1655.2355      1609.5159
#> [57,]      1575.0592      1559.7019      1565.9288      1540.7458
#> [58,]      1693.0000      1693.0000      1693.0000      1693.0000
#> [59,]      1549.6977      1561.8159      1582.6306      1592.3872
#> [60,]       789.0084       838.9124       810.0611       852.3752
#>       Imputation0013 Imputation0014 Imputation0015 Imputation0016
#>  [1,]      1093.8481      1071.4718      1092.7655      1126.2775
#>  [2,]      1624.0000      1624.0000      1624.0000      1624.0000
#>  [3,]      1913.0000      1913.0000      1913.0000      1913.0000
#>  [4,]      1457.1217      1352.5094      1436.0025      1488.1970
#>  [5,]       886.0000       886.0000       886.0000       886.0000
#>  [6,]       582.0446       586.4921       674.6784       661.1517
#>  [7,]      1034.0000      1034.0000      1034.0000      1034.0000
#>  [8,]      1538.3883      1725.3417      1689.3345      1655.1537
#>  [9,]      1630.0000      1630.0000      1630.0000      1630.0000
#> [10,]      1103.0000      1103.0000      1103.0000      1103.0000
#> [11,]      1168.0000      1168.0000      1168.0000      1168.0000
#> [12,]       403.0000       403.0000       403.0000       403.0000
#> [13,]      1099.2447      1171.6622      1152.7267      1078.8667
#> [14,]      1498.0267      1585.4096      1413.0690      1559.6438
#> [15,]      1353.9816      1479.9837      1421.7298      1387.4035
#> [16,]      1235.6822      1192.2196      1170.6093      1254.3678
#> [17,]       992.2841       987.7729       990.3619       983.7958
#> [18,]       394.7057       439.9565       318.7322       389.0587
#> [19,]      1473.8431      1396.6928      1470.9280      1486.9193
#> [20,]      1835.0000      1835.0000      1835.0000      1835.0000
#> [21,]      1584.0000      1584.0000      1584.0000      1584.0000
#> [22,]      1599.2193      1612.6741      1614.9630      1572.3922
#> [23,]       827.5985       908.0868       953.9718       771.5626
#> [24,]       527.8512       478.3003       494.8947       516.1324
#> [25,]      1102.0000      1102.0000      1102.0000      1102.0000
#> [26,]      1669.0000      1669.0000      1669.0000      1669.0000
#> [27,]      1546.8626      1530.7260      1450.1303      1575.5261
#> [28,]      1114.8389      1130.3446      1180.7723      1164.2558
#> [29,]       631.8600       669.7710       625.2794       635.4112
#> [30,]       524.9377       488.3115       524.9627       480.7682
#> [31,]      1592.3384      1572.8313      1553.2591      1556.8885
#> [32,]      1636.7430      1740.2113      1722.0326      1618.1005
#> [33,]      1389.0000      1389.0000      1389.0000      1389.0000
#> [34,]      1144.0000      1144.0000      1144.0000      1144.0000
#> [35,]       741.7205       763.9155       682.2159       711.0152
#> [36,]       483.7742       512.3882       545.3994       502.3419
#> [37,]       876.0000       876.0000       876.0000       876.0000
#> [38,]      1827.0000      1827.0000      1827.0000      1827.0000
#> [39,]      1729.6753      1663.8104      1675.4258      1734.4183
#> [40,]      1696.7297      1683.3075      1611.6183      1772.6084
#> [41,]      1539.0000      1539.0000      1539.0000      1539.0000
#> [42,]       982.0000       982.0000       982.0000       982.0000
#> [43,]       700.6113       746.7333       699.2602       725.2673
#> [44,]      1955.5420      1877.2125      1913.7928      1865.3592
#> [45,]      2218.0000      2218.0000      2218.0000      2218.0000
#> [46,]      1518.4974      1469.1848      1485.7286      1475.5423
#> [47,]      1577.0000      1577.0000      1577.0000      1577.0000
#> [48,]       813.0000       813.0000       813.0000       813.0000
#> [49,]       739.3622       687.5938       706.4320       713.3632
#> [50,]      1529.7910      1551.7005      1526.4612      1478.8805
#> [51,]      1643.8362      1586.1903      1631.8794      1638.9086
#> [52,]      1489.0000      1489.0000      1489.0000      1489.0000
#> [53,]      1088.0000      1088.0000      1088.0000      1088.0000
#> [54,]       745.0000       745.0000       745.0000       745.0000
#> [55,]      1066.9598      1056.3851      1078.9909      1148.5829
#> [56,]      1592.6132      1703.3824      1656.1208      1648.5650
#> [57,]      1519.1220      1521.2087      1541.8943      1597.9675
#> [58,]      1693.0000      1693.0000      1693.0000      1693.0000
#> [59,]      1554.4973      1611.3993      1539.3183      1518.4336
#> [60,]       802.0080       809.7315       755.9797       864.5802
#>       Imputation0017 Imputation0018 Imputation0019
#>  [1,]      1077.0569      1119.5421      1053.9066
#>  [2,]      1624.0000      1624.0000      1624.0000
#>  [3,]      1913.0000      1913.0000      1913.0000
#>  [4,]      1473.6046      1423.4673      1444.3018
#>  [5,]       886.0000       886.0000       886.0000
#>  [6,]       664.5676       631.1631       645.7214
#>  [7,]      1034.0000      1034.0000      1034.0000
#>  [8,]      1441.4168      1635.1539      1658.0938
#>  [9,]      1630.0000      1630.0000      1630.0000
#> [10,]      1103.0000      1103.0000      1103.0000
#> [11,]      1168.0000      1168.0000      1168.0000
#> [12,]       403.0000       403.0000       403.0000
#> [13,]      1128.8063      1117.5227      1041.3750
#> [14,]      1528.6204      1615.5367      1473.4986
#> [15,]      1406.7060      1393.7824      1372.5131
#> [16,]      1258.2359      1212.4757      1267.7986
#> [17,]       961.1616       999.1939       941.6368
#> [18,]       364.8064       420.5530       397.9778
#> [19,]      1399.0642      1437.4562      1423.6587
#> [20,]      1835.0000      1835.0000      1835.0000
#> [21,]      1584.0000      1584.0000      1584.0000
#> [22,]      1556.6737      1551.1836      1578.4124
#> [23,]       926.5545       854.6625       884.1529
#> [24,]       507.5612       563.5588       482.1670
#> [25,]      1102.0000      1102.0000      1102.0000
#> [26,]      1669.0000      1669.0000      1669.0000
#> [27,]      1560.1429      1595.6494      1528.6399
#> [28,]      1158.8471      1176.0471      1163.7739
#> [29,]       612.0004       589.7019       571.0172
#> [30,]       561.5427       517.2779       487.4300
#> [31,]      1530.6469      1593.9960      1542.1033
#> [32,]      1713.7828      1792.6227      1709.1725
#> [33,]      1389.0000      1389.0000      1389.0000
#> [34,]      1144.0000      1144.0000      1144.0000
#> [35,]       683.1011       696.7170       773.2364
#> [36,]       501.9573       530.4440       519.4828
#> [37,]       876.0000       876.0000       876.0000
#> [38,]      1827.0000      1827.0000      1827.0000
#> [39,]      1821.4173      1777.4434      1710.9967
#> [40,]      1667.2483      1781.3283      1714.9404
#> [41,]      1539.0000      1539.0000      1539.0000
#> [42,]       982.0000       982.0000       982.0000
#> [43,]       704.4655       772.7193       778.5592
#> [44,]      1889.1367      1867.3224      1963.8504
#> [45,]      2218.0000      2218.0000      2218.0000
#> [46,]      1500.1001      1501.9545      1443.6153
#> [47,]      1577.0000      1577.0000      1577.0000
#> [48,]       813.0000       813.0000       813.0000
#> [49,]       810.9189       720.5928       655.7400
#> [50,]      1460.1397      1516.4942      1497.1440
#> [51,]      1644.5113      1641.5101      1705.9287
#> [52,]      1489.0000      1489.0000      1489.0000
#> [53,]      1088.0000      1088.0000      1088.0000
#> [54,]       745.0000       745.0000       745.0000
#> [55,]      1136.5998      1087.2059      1120.8691
#> [56,]      1707.3820      1618.9738      1633.9202
#> [57,]      1580.6544      1633.4397      1590.6134
#> [58,]      1693.0000      1693.0000      1693.0000
#> [59,]      1570.0629      1556.7758      1602.0193
#> [60,]       802.1174       721.3224       788.5303
#>