Deprecated functions

These functions will be removed from multimput in the future.

generateData(
  intercept = 2,
  n.year = 24,
  n.period = 6,
  n.site = 20,
  year.factor = FALSE,
  period.factor = FALSE,
  site.factor = FALSE,
  trend = 0.01,
  sd.rw.year = 0.1,
  amplitude.period = 1,
  mean.phase.period = 0,
  sd.phase.period = 0.2,
  sd.site = 1,
  sd.rw.site = 0.02,
  sd.noise = 0.01,
  size = 2,
  n.run = 10,
  as.list = FALSE,
  details = FALSE
)

missingAtRandom(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed"
)

missingCurrentCount(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed"
)

missingObserved(
  dataset,
  count.variable = "Count",
  observed.variable = "Observed",
  site.variable = "Site",
  year.variable = "Year",
  period.variable = "Period"
)

missingVolunteer(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed",
  year.variable = "Year",
  site.variable = "Site",
  max.count = 100
)

Arguments

intercept

The global mean on the log-scale.

n.year

The number of years.

n.period

The number of periods.

n.site

The number of sites.

year.factor

Convert year to a factor. Defaults to FALSE.

period.factor

Convert period to a factor. Defaults to FALSE.

site.factor

Convert site to a factor. Defaults to FALSE.

trend

The long-term linear trend on the log-scale.

sd.rw.year

The standard deviation of the year effects on the log-scale.

amplitude.period

The amplitude of the periodic effect on the log-scale.

mean.phase.period

The mean of the phase of the periodic effect among years. Defaults to 0.

sd.phase.period

The standard deviation of the phase of the periodic effect among years.

sd.site

The standard deviation of the site effects on the log-scale.

sd.rw.site

The standard deviation of the random walk along year per site on the log-scale.

sd.noise

The standard deviation of the noise effects on the log-scale.

size

The size parameter of the negative binomial distribution.

n.run

The number of runs with the same mu.

as.list

Return the dataset as a list rather than a data.frame. Defaults to FALSE.

details

Add variables containing the year, period and site effects. Defaults tot FALSE.

dataset

A dataset to a the observation with missing data.

proportion

The proportion of observations that will be missing.

count.variable

The name of the variable holding the counts.

observed.variable

The name of the variable holding the observed values = either count or missing.

site.variable

The name of the variable holding the sites.

year.variable

The name of the variable holding the years.

period.variable

The name of the variable holding the period.

max.count

The maximum count.