Aggregate an imputed dataset

Aggregate an imputed dataset

aggregate_impute(object, grouping, fun, filter, join)

# S4 method for ANY
aggregate_impute(object, grouping, fun, filter, join)

# S4 method for rawImputed
aggregate_impute(object, grouping, fun, filter, join)

# S4 method for aggregatedImputed
aggregate_impute(object, grouping, fun, filter,
  join)

Arguments

object
a rawImputed object

grouping
a vector of variables names to group the aggregation on

fun
the function to aggregate

filter
an optional argument to filter the raw dataset before aggregation. Will be passed to the .dots argument of filter_

join
an optional argument to filter the raw dataset based on a data.frame. A semi_join will be applied with join or each element of join in case join is a list

Examples

dataset <- generateData(n.year = 10, n.site = 50, n.run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)

#> An object of class "aggregatedImputed"
#> Slot "Covariate":
#>    Year Period
#> 1     1      1
#> 2     1      2
#> 3     1      3
#> 4     1      4
#> 5     1      5
#> 6     1      6
#> 7     2      1
#> 8     2      2
#> 9     2      3
#> 10    2      4
#> 11    2      5
#> 12    2      6
#> 13    3      1
#> 14    3      2
#> 15    3      3
#> 16    3      4
#> 17    3      5
#> 18    3      6
#> 19    4      1
#> 20    4      2
#> 21    4      3
#> 22    4      4
#> 23    4      5
#> 24    4      6
#> 25    5      1
#> 26    5      2
#> 27    5      3
#> 28    5      4
#> 29    5      5
#> 30    5      6
#> 31    6      1
#> 32    6      2
#> 33    6      3
#> 34    6      4
#> 35    6      5
#> 36    6      6
#> 37    7      1
#> 38    7      2
#> 39    7      3
#> 40    7      4
#> 41    7      5
#> 42    7      6
#> 43    8      1
#> 44    8      2
#> 45    8      3
#> 46    8      4
#> 47    8      5
#> 48    8      6
#> 49    9      1
#> 50    9      2
#> 51    9      3
#> 52    9      4
#> 53    9      5
#> 54    9      6
#> 55   10      1
#> 56   10      2
#> 57   10      3
#> 58   10      4
#> 59   10      5
#> 60   10      6
#> 
#> Slot "Imputation":
#>       Imputation0001 Imputation0002 Imputation0003 Imputation0004
#>  [1,]       718.5883       749.9646       649.4030       744.6085
#>  [2,]      1028.4235      1074.2759      1114.1151      1068.5431
#>  [3,]      1966.8737      1980.5191      1980.4241      2002.9352
#>  [4,]      1307.0000      1307.0000      1307.0000      1307.0000
#>  [5,]       926.0000       926.0000       926.0000       926.0000
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       735.7928       816.5656       716.7535       753.9461
#>  [8,]      1059.9920      1101.6022      1106.6852      1052.0947
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1481.4096      1532.6598      1517.6019      1497.9696
#> [11,]      1021.0000      1021.0000      1021.0000      1021.0000
#> [12,]       694.6790       612.3814       616.7732       705.0666
#> [13,]      1000.2983       998.4289      1003.3064       956.0384
#> [14,]      1458.5238      1459.1888      1419.8665      1477.9573
#> [15,]      1137.9126      1093.2326      1147.8696      1100.4572
#> [16,]       976.5884      1017.2453       939.6257      1011.9500
#> [17,]       410.4408       430.0155       444.6435       472.3693
#> [18,]       381.4068       337.5996       399.9656       431.3506
#> [19,]       461.1316       580.5729       526.0355       560.6623
#> [20,]      1135.1194      1173.8457      1180.6608      1196.4076
#> [21,]      1112.4984      1084.0128      1120.7308      1121.0417
#> [22,]      1286.0000      1286.0000      1286.0000      1286.0000
#> [23,]      1254.9393      1292.4487      1195.0765      1266.8756
#> [24,]       532.0000       532.0000       532.0000       532.0000
#> [25,]      1049.0000      1049.0000      1049.0000      1049.0000
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1341.0428      1356.6675      1378.2020      1371.8330
#> [28,]      1174.0702      1154.5532      1167.3949      1114.9617
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       614.0000       614.0000       614.0000       614.0000
#> [31,]       811.6428       824.1896       807.1196       737.9303
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1821.8398      1877.0443      1822.5616      1853.1743
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1437.0000      1437.0000      1437.0000      1437.0000
#> [36,]       663.0000       663.0000       663.0000       663.0000
#> [37,]       841.6259       796.9614       826.3565       806.9893
#> [38,]      1367.2643      1297.9040      1384.9704      1374.0046
#> [39,]      1195.2631      1188.0141      1115.2038      1146.4023
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       811.0000       811.0000       811.0000       811.0000
#> [42,]       808.3043       780.4581       868.9186       875.6635
#> [43,]       566.4306       621.3091       658.4122       606.1020
#> [44,]      1040.8439      1039.4629       994.9902      1010.6075
#> [45,]      1098.0000      1098.0000      1098.0000      1098.0000
#> [46,]      1078.0000      1078.0000      1078.0000      1078.0000
#> [47,]       869.6535       899.8814       841.6425       833.1229
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       661.5423       648.7767       603.1757       596.1221
#> [50,]      1459.3303      1448.4491      1425.6352      1463.2993
#> [51,]      1426.1164      1372.2517      1364.1325      1347.7357
#> [52,]      1165.2020      1083.5051      1140.1290      1102.4406
#> [53,]       652.0000       652.0000       652.0000       652.0000
#> [54,]       339.7917       324.0303       365.6612       381.0228
#> [55,]       666.5804       665.1697       679.7950       618.4251
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1335.0655      1342.3844      1300.9850      1398.9886
#> [58,]      1526.6935      1544.7977      1527.7053      1559.1870
#> [59,]       820.5133       695.7398       720.6814       798.2930
#> [60,]       745.0719       704.3887       723.2263       809.8359
#>       Imputation0005 Imputation0006 Imputation0007 Imputation0008
#>  [1,]       719.8979       820.9679       711.4112       735.0291
#>  [2,]      1058.7651      1122.9956      1124.9606      1087.7879
#>  [3,]      1967.4173      1970.2536      1959.7723      1932.2803
#>  [4,]      1307.0000      1307.0000      1307.0000      1307.0000
#>  [5,]       926.0000       926.0000       926.0000       926.0000
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       773.2273       732.0139       810.9562       751.6768
#>  [8,]      1133.6618      1107.2670      1093.9742      1086.5489
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1511.9963      1538.6161      1540.6267      1519.9075
#> [11,]      1021.0000      1021.0000      1021.0000      1021.0000
#> [12,]       646.4205       695.5316       626.6582       643.2031
#> [13,]      1024.5651       961.8703      1052.0599      1042.7600
#> [14,]      1476.0822      1438.3406      1492.3919      1455.0738
#> [15,]      1131.6507      1190.8939      1201.6138      1040.8174
#> [16,]      1002.8890       991.6624       964.0728       982.3806
#> [17,]       430.9135       434.9796       434.4371       439.7413
#> [18,]       423.2493       436.8049       384.7612       369.3981
#> [19,]       491.8624       487.9551       499.6735       523.3449
#> [20,]      1163.4330      1157.5887      1153.9439      1160.1675
#> [21,]      1088.4471      1166.1609      1106.3912      1097.6209
#> [22,]      1286.0000      1286.0000      1286.0000      1286.0000
#> [23,]      1210.7059      1300.9177      1172.7626      1281.8346
#> [24,]       532.0000       532.0000       532.0000       532.0000
#> [25,]      1049.0000      1049.0000      1049.0000      1049.0000
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1351.8140      1371.0425      1323.9390      1348.9054
#> [28,]      1162.9369      1177.5030      1205.3258      1182.2612
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       614.0000       614.0000       614.0000       614.0000
#> [31,]       786.7079       777.9779       751.5731       823.2278
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1840.0916      1859.8535      1836.7375      1846.5267
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1437.0000      1437.0000      1437.0000      1437.0000
#> [36,]       663.0000       663.0000       663.0000       663.0000
#> [37,]       769.0886       790.6803       824.4296       846.1503
#> [38,]      1306.3577      1374.6806      1348.7546      1329.7961
#> [39,]      1094.2698      1140.0101      1176.6248      1142.2331
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       811.0000       811.0000       811.0000       811.0000
#> [42,]       856.5463       795.6222       887.5470       865.3100
#> [43,]       583.0238       623.6121       636.6843       640.0364
#> [44,]      1018.3765      1004.7889      1037.9686      1061.2209
#> [45,]      1098.0000      1098.0000      1098.0000      1098.0000
#> [46,]      1078.0000      1078.0000      1078.0000      1078.0000
#> [47,]       863.5958       879.9095       847.0642       909.9704
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       624.7712       568.7437       622.0291       714.9076
#> [50,]      1451.3662      1498.1847      1452.6254      1500.7518
#> [51,]      1407.1627      1393.6893      1345.2557      1330.8270
#> [52,]      1111.6450      1117.6136      1152.0673      1164.0446
#> [53,]       652.0000       652.0000       652.0000       652.0000
#> [54,]       365.6739       462.0058       430.2740       391.7001
#> [55,]       627.9065       691.9712       685.8229       635.6011
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1417.5665      1313.2052      1339.7499      1361.6093
#> [58,]      1573.2665      1572.5968      1614.7141      1572.5446
#> [59,]       758.5502       733.7630       743.4620       803.9697
#> [60,]       703.5300       807.3775       796.2987       742.1737
#>       Imputation0009 Imputation0010 Imputation0011 Imputation0012
#>  [1,]       707.2696       755.9886       743.8176       743.5523
#>  [2,]      1007.5876      1050.6807      1157.6389      1092.7375
#>  [3,]      1941.0054      2000.7355      2008.8930      1951.0680
#>  [4,]      1307.0000      1307.0000      1307.0000      1307.0000
#>  [5,]       926.0000       926.0000       926.0000       926.0000
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       754.6754       800.1477       811.5103       762.1837
#>  [8,]      1103.5860      1075.5376      1084.8800      1088.4558
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1465.2498      1518.2215      1426.1852      1557.5612
#> [11,]      1021.0000      1021.0000      1021.0000      1021.0000
#> [12,]       674.6905       623.8096       671.2194       609.6190
#> [13,]       950.8623      1045.8564       978.9639       964.1508
#> [14,]      1464.0289      1431.8895      1456.1176      1423.4834
#> [15,]      1096.7573      1111.9509      1138.7715      1108.8280
#> [16,]      1014.2389       958.4330       986.2633       927.1639
#> [17,]       413.7847       462.0055       461.3782       434.0564
#> [18,]       423.1388       396.7636       382.4401       389.2998
#> [19,]       481.0893       486.1772       473.4713       442.3815
#> [20,]      1167.9031      1158.7539      1146.5742      1160.5048
#> [21,]      1040.2401      1110.5314      1154.2810      1062.6375
#> [22,]      1286.0000      1286.0000      1286.0000      1286.0000
#> [23,]      1266.2705      1239.1381      1304.8740      1242.2293
#> [24,]       532.0000       532.0000       532.0000       532.0000
#> [25,]      1049.0000      1049.0000      1049.0000      1049.0000
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1349.9414      1320.7601      1358.4890      1371.1903
#> [28,]      1123.1285      1182.7369      1173.1586      1189.8383
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       614.0000       614.0000       614.0000       614.0000
#> [31,]       768.2278       832.2681       804.5233       805.0342
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1825.6181      1821.6614      1875.0070      1815.1442
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1437.0000      1437.0000      1437.0000      1437.0000
#> [36,]       663.0000       663.0000       663.0000       663.0000
#> [37,]       802.6840       800.3707       793.2568       780.6607
#> [38,]      1353.0613      1309.9330      1350.0367      1340.9836
#> [39,]      1087.5537      1070.6352      1166.1670      1153.5202
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       811.0000       811.0000       811.0000       811.0000
#> [42,]       924.9508       828.1355       841.0784       864.6941
#> [43,]       572.6672       623.9429       610.2361       584.1224
#> [44,]      1074.3458      1062.5008      1057.2899      1047.3941
#> [45,]      1098.0000      1098.0000      1098.0000      1098.0000
#> [46,]      1078.0000      1078.0000      1078.0000      1078.0000
#> [47,]       887.1846       858.8135       826.0497       858.5676
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       687.0952       629.7158       590.1435       599.0114
#> [50,]      1492.2561      1458.4210      1460.7141      1426.4963
#> [51,]      1410.0840      1363.1988      1478.4857      1385.6794
#> [52,]      1130.4498      1094.7466      1115.3655      1093.1347
#> [53,]       652.0000       652.0000       652.0000       652.0000
#> [54,]       328.2961       364.0819       295.3250       317.3210
#> [55,]       651.9878       670.3537       618.2700       675.5003
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1423.1736      1340.6304      1337.3948      1308.2359
#> [58,]      1526.6746      1577.2359      1576.7668      1565.7060
#> [59,]       671.3602       733.5386       811.5949       788.3018
#> [60,]       866.2835       693.6386       697.9953       703.1767
#>       Imputation0013 Imputation0014 Imputation0015 Imputation0016
#>  [1,]       720.3538       735.3544       724.7812       800.4506
#>  [2,]      1111.3009      1152.0313      1056.0215      1042.4599
#>  [3,]      1972.0881      1970.0750      1938.2641      2033.4175
#>  [4,]      1307.0000      1307.0000      1307.0000      1307.0000
#>  [5,]       926.0000       926.0000       926.0000       926.0000
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       751.1938       779.8528       760.5213       734.6595
#>  [8,]      1058.6024      1148.2579      1110.6431      1010.2967
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1487.3992      1524.7393      1479.7805      1491.2718
#> [11,]      1021.0000      1021.0000      1021.0000      1021.0000
#> [12,]       667.0251       612.7977       633.1297       728.1211
#> [13,]       965.2260      1005.2693       985.3457       936.6037
#> [14,]      1471.1116      1454.8222      1460.9565      1496.3467
#> [15,]      1102.9137      1115.4968      1091.2336      1115.4871
#> [16,]       955.7310       969.8468       979.8819       954.0636
#> [17,]       425.9498       386.8645       417.4505       388.1906
#> [18,]       348.3178       371.2270       389.1306       414.5135
#> [19,]       470.9341       532.5016       499.2576       522.6829
#> [20,]      1149.7078      1088.4654      1120.9856      1150.6977
#> [21,]      1145.2527      1177.1734      1200.3354      1102.5259
#> [22,]      1286.0000      1286.0000      1286.0000      1286.0000
#> [23,]      1305.3034      1292.3918      1191.0559      1224.7438
#> [24,]       532.0000       532.0000       532.0000       532.0000
#> [25,]      1049.0000      1049.0000      1049.0000      1049.0000
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1307.2261      1306.1348      1283.2340      1357.9586
#> [28,]      1190.0409      1181.0572      1159.1851      1110.5316
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       614.0000       614.0000       614.0000       614.0000
#> [31,]       848.3246       797.2647       758.3672       763.4124
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1857.6430      1824.9737      1863.3329      1776.1526
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1437.0000      1437.0000      1437.0000      1437.0000
#> [36,]       663.0000       663.0000       663.0000       663.0000
#> [37,]       834.3440       791.7454       766.0877       837.5136
#> [38,]      1345.4532      1378.5705      1348.6069      1430.1611
#> [39,]      1178.1544      1127.3082      1172.7116      1180.9174
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       811.0000       811.0000       811.0000       811.0000
#> [42,]       870.4352       880.8405       887.8295       924.0474
#> [43,]       656.1372       623.0834       646.7005       637.4147
#> [44,]       986.5911       976.9344       989.9967      1020.5312
#> [45,]      1098.0000      1098.0000      1098.0000      1098.0000
#> [46,]      1078.0000      1078.0000      1078.0000      1078.0000
#> [47,]       879.5522       892.4130       843.7168       870.4020
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       643.7706       607.3797       636.7068       737.1356
#> [50,]      1474.1017      1482.2389      1444.8021      1432.0916
#> [51,]      1424.1883      1318.0505      1299.7008      1369.5569
#> [52,]      1095.5260      1078.4368      1088.3643      1120.7511
#> [53,]       652.0000       652.0000       652.0000       652.0000
#> [54,]       427.2101       380.7207       348.2064       360.2142
#> [55,]       590.9640       657.5206       637.8211       609.9627
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1368.0130      1433.9032      1364.3047      1324.9288
#> [58,]      1551.5226      1559.9088      1534.8536      1579.5389
#> [59,]       720.0410       808.5336       764.2141       757.6647
#> [60,]       728.2197       692.3014       697.8832       778.7676
#>       Imputation0017 Imputation0018 Imputation0019
#>  [1,]       726.6541       717.1503       693.1079
#>  [2,]      1118.6779      1124.3746      1029.3342
#>  [3,]      1957.2410      2003.9829      1925.6649
#>  [4,]      1307.0000      1307.0000      1307.0000
#>  [5,]       926.0000       926.0000       926.0000
#>  [6,]       697.0000       697.0000       697.0000
#>  [7,]       779.8125       787.6784       774.7176
#>  [8,]      1114.9591      1097.1343      1096.6327
#>  [9,]      1821.0000      1821.0000      1821.0000
#> [10,]      1490.0523      1493.7758      1510.5673
#> [11,]      1021.0000      1021.0000      1021.0000
#> [12,]       651.0110       735.6645       674.2265
#> [13,]       981.1213       947.5423      1047.2512
#> [14,]      1449.3971      1461.2274      1501.4204
#> [15,]      1121.2963      1092.9916      1080.2189
#> [16,]       993.1823      1033.7405       977.6759
#> [17,]       403.3062       420.1954       482.7992
#> [18,]       399.6397       355.0418       369.3346
#> [19,]       586.9854       469.2676       523.5779
#> [20,]      1190.4884      1150.0457      1164.4119
#> [21,]      1158.7956      1079.3201      1102.4440
#> [22,]      1286.0000      1286.0000      1286.0000
#> [23,]      1240.8929      1208.2081      1207.0073
#> [24,]       532.0000       532.0000       532.0000
#> [25,]      1049.0000      1049.0000      1049.0000
#> [26,]      1746.0000      1746.0000      1746.0000
#> [27,]      1275.8957      1343.0702      1352.0896
#> [28,]      1166.1947      1095.3719      1183.8164
#> [29,]       584.0000       584.0000       584.0000
#> [30,]       614.0000       614.0000       614.0000
#> [31,]       756.9912       801.7520       812.6427
#> [32,]      1291.0000      1291.0000      1291.0000
#> [33,]      1859.3247      1850.3077      1857.3433
#> [34,]      1832.0000      1832.0000      1832.0000
#> [35,]      1437.0000      1437.0000      1437.0000
#> [36,]       663.0000       663.0000       663.0000
#> [37,]       868.6450       816.7613       835.6662
#> [38,]      1342.9432      1362.0304      1404.4345
#> [39,]      1172.5180      1141.9985      1152.5931
#> [40,]      1200.0000      1200.0000      1200.0000
#> [41,]       811.0000       811.0000       811.0000
#> [42,]       937.8485       756.9865       880.6206
#> [43,]       606.1581       650.5388       660.5647
#> [44,]       994.4879      1002.7427      1023.9431
#> [45,]      1098.0000      1098.0000      1098.0000
#> [46,]      1078.0000      1078.0000      1078.0000
#> [47,]       846.9062       907.5552       843.9757
#> [48,]       526.0000       526.0000       526.0000
#> [49,]       620.1510       601.8352       603.9513
#> [50,]      1459.0669      1501.8272      1434.5357
#> [51,]      1463.5051      1360.2141      1390.1230
#> [52,]      1125.0446      1134.1575      1129.2019
#> [53,]       652.0000       652.0000       652.0000
#> [54,]       379.4620       331.8954       421.4041
#> [55,]       664.7569       663.0488       660.0714
#> [56,]       996.0000       996.0000       996.0000
#> [57,]      1295.9403      1246.3752      1343.6853
#> [58,]      1529.5896      1574.8787      1552.0884
#> [59,]       724.7221       772.9462       792.1869
#> [60,]       697.3817       758.7513       757.3905
#>