Aggregate an imputed dataset

Aggregate an imputed dataset

Usage

aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for class 'ANY'
aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for class 'rawImputed'
aggregate_impute(object, grouping, fun, filter = list(), join)

# S4 method for class 'aggregatedImputed'
aggregate_impute(object, grouping, fun, filter = list(), join)

Arguments

object

A rawImputed object.

grouping

A vector of variables names to group the aggregation on.

fun

The function to aggregate.

filter

An optional argument to filter the raw dataset before aggregation. Will be passed to dplyr::filter().

join

An optional argument to filter the raw dataset based on a data.frame. A dplyr::semi_join() will be applied with join or each element of join in case join is a list.

Examples

dataset <- generate_data(n_year = 10, n_site = 50, n_run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
#> An object of class "aggregatedImputed"
#> Slot "Covariate":
#>    Year Period
#> 1     1      1
#> 2     1      2
#> 3     1      3
#> 4     1      4
#> 5     1      5
#> 6     1      6
#> 7     2      1
#> 8     2      2
#> 9     2      3
#> 10    2      4
#> 11    2      5
#> 12    2      6
#> 13    3      1
#> 14    3      2
#> 15    3      3
#> 16    3      4
#> 17    3      5
#> 18    3      6
#> 19    4      1
#> 20    4      2
#> 21    4      3
#> 22    4      4
#> 23    4      5
#> 24    4      6
#> 25    5      1
#> 26    5      2
#> 27    5      3
#> 28    5      4
#> 29    5      5
#> 30    5      6
#> 31    6      1
#> 32    6      2
#> 33    6      3
#> 34    6      4
#> 35    6      5
#> 36    6      6
#> 37    7      1
#> 38    7      2
#> 39    7      3
#> 40    7      4
#> 41    7      5
#> 42    7      6
#> 43    8      1
#> 44    8      2
#> 45    8      3
#> 46    8      4
#> 47    8      5
#> 48    8      6
#> 49    9      1
#> 50    9      2
#> 51    9      3
#> 52    9      4
#> 53    9      5
#> 54    9      6
#> 55   10      1
#> 56   10      2
#> 57   10      3
#> 58   10      4
#> 59   10      5
#> 60   10      6
#> 
#> Slot "Imputation":
#>       Imputation0001 Imputation0002 Imputation0003 Imputation0004
#>  [1,]       756.3526       734.4290       734.7943       754.6791
#>  [2,]      1067.0000      1067.0000      1067.0000      1067.0000
#>  [3,]      1947.1719      1962.5697      1969.5826      2025.1728
#>  [4,]      1300.8334      1314.5699      1314.4742      1337.1357
#>  [5,]       923.6729       878.4131       940.3766       952.5347
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       674.9876       843.9430       732.4395       701.7702
#>  [8,]      1096.7673      1103.7887      1044.7352      1112.1938
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1514.3596      1544.7936      1486.1576      1477.5799
#> [11,]      1003.2117      1018.3315      1072.5774      1079.9879
#> [12,]       620.6377       697.9800       578.5839       664.1148
#> [13,]      1007.5702      1001.5890       881.7306      1036.8691
#> [14,]      1445.5250      1463.4137      1455.3599      1408.6289
#> [15,]      1116.7524      1157.6726      1079.5502      1152.3430
#> [16,]       965.0000       965.0000       965.0000       965.0000
#> [17,]       429.0955       427.1248       421.8147       430.7142
#> [18,]       387.0000       387.0000       387.0000       387.0000
#> [19,]       499.2491       550.8225       535.6697       515.9135
#> [20,]      1137.0000      1137.0000      1137.0000      1137.0000
#> [21,]      1080.4678      1098.6832      1081.4857      1113.1610
#> [22,]      1240.1549      1229.2085      1206.2580      1244.1476
#> [23,]      1229.9433      1228.0616      1232.9711      1185.3930
#> [24,]       548.2974       477.2315       501.8332       493.6696
#> [25,]      1085.6475      1040.7069      1070.2838      1050.7967
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1334.0000      1334.0000      1334.0000      1334.0000
#> [28,]      1162.8787      1135.8472      1198.9784      1161.8557
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       575.5428       642.2964       566.1852       649.5908
#> [31,]       785.0000       785.0000       785.0000       785.0000
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1912.0000      1912.0000      1912.0000      1912.0000
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1497.8362      1509.0415      1506.5780      1375.2277
#> [36,]       689.7552       670.1139       683.0374       630.2704
#> [37,]       824.4284       855.9830       754.8499       850.5965
#> [38,]      1403.7285      1321.7283      1370.7881      1436.1340
#> [39,]      1125.0000      1125.0000      1125.0000      1125.0000
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       819.9588       836.8221       804.0500       795.7137
#> [42,]       879.1477       861.7562       876.6918       872.0062
#> [43,]       621.0000       621.0000       621.0000       621.0000
#> [44,]       991.1248      1030.1170      1036.9789      1052.8339
#> [45,]      1073.1094      1146.0057      1136.9600      1028.7495
#> [46,]      1080.4544      1101.8357      1088.5774      1075.5436
#> [47,]       875.0000       875.0000       875.0000       875.0000
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       520.4784       513.8523       401.0042       490.8617
#> [50,]      1496.4838      1512.2244      1559.3394      1540.6975
#> [51,]      1363.0000      1363.0000      1363.0000      1363.0000
#> [52,]      1132.3621      1088.2788      1151.0379      1182.6207
#> [53,]       621.3340       611.5562       618.5966       670.3268
#> [54,]       405.0000       405.0000       405.0000       405.0000
#> [55,]       658.0000       658.0000       658.0000       658.0000
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1382.9942      1333.6401      1366.9134      1402.3085
#> [58,]      1548.0000      1548.0000      1548.0000      1548.0000
#> [59,]       772.0000       772.0000       772.0000       772.0000
#> [60,]       567.5417       619.2688       646.8944       646.2323
#>       Imputation0005 Imputation0006 Imputation0007 Imputation0008
#>  [1,]       777.1575       777.0502       725.7339       782.4002
#>  [2,]      1067.0000      1067.0000      1067.0000      1067.0000
#>  [3,]      1940.5955      1909.9186      1956.6923      2038.3129
#>  [4,]      1301.3806      1304.2358      1293.6845      1266.0089
#>  [5,]       918.5958       952.7087       983.9895       931.2888
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       789.9625       722.3787       782.7494       746.5058
#>  [8,]      1106.4081      1102.4398      1101.4732      1151.2396
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1508.2606      1524.6855      1491.6162      1554.9514
#> [11,]      1027.2552      1073.4629      1088.5658      1022.3842
#> [12,]       620.6114       649.9751       615.9351       663.8915
#> [13,]       981.7254       935.6395       948.1448      1000.8034
#> [14,]      1435.9619      1404.1054      1424.6858      1458.3267
#> [15,]      1143.2234      1131.9240      1104.1557      1122.5821
#> [16,]       965.0000       965.0000       965.0000       965.0000
#> [17,]       404.3039       450.7239       467.6272       391.7591
#> [18,]       387.0000       387.0000       387.0000       387.0000
#> [19,]       530.0287       556.8164       558.8397       537.9898
#> [20,]      1137.0000      1137.0000      1137.0000      1137.0000
#> [21,]      1127.3270      1126.6531      1169.0293      1126.6007
#> [22,]      1232.1431      1279.2419      1233.4098      1281.8244
#> [23,]      1254.3693      1191.2632      1282.0444      1272.6835
#> [24,]       575.4052       580.3917       520.2302       494.4040
#> [25,]      1012.6618      1034.3870      1068.3449      1090.1999
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1334.0000      1334.0000      1334.0000      1334.0000
#> [28,]      1147.8557      1206.0343      1190.0132      1214.1559
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       634.9835       693.5978       654.4297       601.3246
#> [31,]       785.0000       785.0000       785.0000       785.0000
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1912.0000      1912.0000      1912.0000      1912.0000
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1433.8443      1489.5043      1456.7516      1478.3300
#> [36,]       678.5509       693.2098       721.2096       697.9982
#> [37,]       825.7455       927.3899       817.2105       840.9626
#> [38,]      1310.3743      1481.5412      1348.5855      1363.7032
#> [39,]      1125.0000      1125.0000      1125.0000      1125.0000
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       773.7779       755.2333       766.6573       777.9643
#> [42,]       914.6413       842.1325       888.4747       905.2454
#> [43,]       621.0000       621.0000       621.0000       621.0000
#> [44,]      1019.6329      1013.7484      1010.0786      1016.3449
#> [45,]      1153.7357      1145.5965      1119.0917      1140.4010
#> [46,]      1136.9768      1042.0315      1103.8561      1114.8815
#> [47,]       875.0000       875.0000       875.0000       875.0000
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       436.5086       444.5606       535.8038       463.6371
#> [50,]      1553.2292      1561.6815      1515.1499      1552.6508
#> [51,]      1363.0000      1363.0000      1363.0000      1363.0000
#> [52,]      1174.4683      1188.1093      1135.7376      1120.2778
#> [53,]       662.5482       623.0870       626.0436       624.0827
#> [54,]       405.0000       405.0000       405.0000       405.0000
#> [55,]       658.0000       658.0000       658.0000       658.0000
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1308.9352      1403.7359      1426.9446      1346.0950
#> [58,]      1548.0000      1548.0000      1548.0000      1548.0000
#> [59,]       772.0000       772.0000       772.0000       772.0000
#> [60,]       613.0373       560.2690       609.4566       646.2113
#>       Imputation0009 Imputation0010 Imputation0011 Imputation0012
#>  [1,]       762.3365       702.4994       754.5193       701.4686
#>  [2,]      1067.0000      1067.0000      1067.0000      1067.0000
#>  [3,]      1826.9642      1965.3479      2002.8980      1935.5162
#>  [4,]      1274.7922      1334.9214      1343.1334      1284.9220
#>  [5,]       987.0087       923.4516       898.0379       901.0275
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       706.1038       792.2479       794.5704       749.4217
#>  [8,]      1184.2485      1145.9764      1115.2753      1138.1598
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1532.0102      1503.4456      1470.4584      1503.1981
#> [11,]      1008.6793      1030.2800      1086.4156      1025.6837
#> [12,]       637.1983       641.4424       672.9571       675.4227
#> [13,]      1041.2325      1002.3852       967.5405      1003.9373
#> [14,]      1460.2299      1435.9974      1436.7831      1406.8152
#> [15,]      1154.6468      1098.4794      1126.4899      1067.0077
#> [16,]       965.0000       965.0000       965.0000       965.0000
#> [17,]       403.4692       405.0092       455.8890       394.5001
#> [18,]       387.0000       387.0000       387.0000       387.0000
#> [19,]       482.9874       536.2932       443.6764       575.8810
#> [20,]      1137.0000      1137.0000      1137.0000      1137.0000
#> [21,]      1080.4487      1131.3208      1130.8488      1119.7201
#> [22,]      1273.2778      1239.2402      1241.5470      1207.1243
#> [23,]      1180.1830      1275.8002      1208.4689      1193.5587
#> [24,]       570.5918       551.7186       590.6041       527.2699
#> [25,]      1046.4649      1044.1372      1036.9793      1024.3054
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1334.0000      1334.0000      1334.0000      1334.0000
#> [28,]      1120.7270      1210.7327      1157.9030      1171.4021
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       609.9457       551.9321       589.3027       537.6691
#> [31,]       785.0000       785.0000       785.0000       785.0000
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1912.0000      1912.0000      1912.0000      1912.0000
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1439.4738      1522.3846      1442.0762      1500.1594
#> [36,]       638.4892       698.4770       688.8377       705.6235
#> [37,]       813.0453       862.0413       849.8011       849.5343
#> [38,]      1295.8157      1409.4346      1401.1996      1251.0787
#> [39,]      1125.0000      1125.0000      1125.0000      1125.0000
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       781.5310       753.2245       830.3501       800.2953
#> [42,]       956.3875       851.1245       886.6091       823.3506
#> [43,]       621.0000       621.0000       621.0000       621.0000
#> [44,]      1024.1337      1014.9217      1002.6583      1016.6846
#> [45,]      1090.5448      1084.0804      1144.9754      1144.0011
#> [46,]      1083.3307      1117.9794      1052.4657      1138.1113
#> [47,]       875.0000       875.0000       875.0000       875.0000
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       417.6197       368.2496       488.7746       443.2056
#> [50,]      1494.1742      1492.3820      1485.8659      1535.0662
#> [51,]      1363.0000      1363.0000      1363.0000      1363.0000
#> [52,]      1174.3571      1147.8157      1133.4019      1140.3049
#> [53,]       639.7818       594.2266       693.6094       662.0018
#> [54,]       405.0000       405.0000       405.0000       405.0000
#> [55,]       658.0000       658.0000       658.0000       658.0000
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1451.1800      1303.8054      1333.6563      1306.0203
#> [58,]      1548.0000      1548.0000      1548.0000      1548.0000
#> [59,]       772.0000       772.0000       772.0000       772.0000
#> [60,]       618.1451       565.1245       617.9595       625.1172
#>       Imputation0013 Imputation0014 Imputation0015 Imputation0016
#>  [1,]       734.9059       682.9614       726.0731       722.4775
#>  [2,]      1067.0000      1067.0000      1067.0000      1067.0000
#>  [3,]      1965.6300      2050.8871      2018.6023      2003.7769
#>  [4,]      1306.0825      1304.0560      1272.0326      1367.8217
#>  [5,]       943.7690       929.8961       903.8432       961.4262
#>  [6,]       697.0000       697.0000       697.0000       697.0000
#>  [7,]       771.7014       770.3811       833.0295       719.1908
#>  [8,]      1052.5544      1020.4598      1057.9895      1105.6142
#>  [9,]      1821.0000      1821.0000      1821.0000      1821.0000
#> [10,]      1524.3258      1537.2743      1488.2460      1515.1131
#> [11,]      1107.7228      1080.8152      1042.2053      1014.4154
#> [12,]       653.2549       667.7424       612.3179       674.6549
#> [13,]       983.5168       961.9835       933.9022       935.1612
#> [14,]      1415.8872      1398.7146      1459.7217      1506.5152
#> [15,]      1095.7598      1109.9671      1120.0672      1094.0817
#> [16,]       965.0000       965.0000       965.0000       965.0000
#> [17,]       419.2068       400.8527       411.1601       402.4652
#> [18,]       387.0000       387.0000       387.0000       387.0000
#> [19,]       505.2765       542.8521       497.6097       509.1735
#> [20,]      1137.0000      1137.0000      1137.0000      1137.0000
#> [21,]      1105.4495      1113.8871      1088.6780      1133.6379
#> [22,]      1255.0148      1263.2007      1225.5397      1212.7531
#> [23,]      1194.6409      1234.9469      1214.8926      1165.8309
#> [24,]       525.2650       510.5291       496.1009       542.9157
#> [25,]      1078.3206      1035.4586      1009.6423      1081.5098
#> [26,]      1746.0000      1746.0000      1746.0000      1746.0000
#> [27,]      1334.0000      1334.0000      1334.0000      1334.0000
#> [28,]      1168.0524      1162.3855      1154.7688      1181.2978
#> [29,]       584.0000       584.0000       584.0000       584.0000
#> [30,]       694.3590       620.3692       633.8897       531.6744
#> [31,]       785.0000       785.0000       785.0000       785.0000
#> [32,]      1291.0000      1291.0000      1291.0000      1291.0000
#> [33,]      1912.0000      1912.0000      1912.0000      1912.0000
#> [34,]      1832.0000      1832.0000      1832.0000      1832.0000
#> [35,]      1458.6805      1474.2558      1415.3017      1392.3547
#> [36,]       705.8275       696.7865       674.7753       625.8122
#> [37,]       826.2039       841.2898       830.6565       906.7560
#> [38,]      1390.1662      1357.6213      1430.8220      1330.0074
#> [39,]      1125.0000      1125.0000      1125.0000      1125.0000
#> [40,]      1200.0000      1200.0000      1200.0000      1200.0000
#> [41,]       841.0511       748.9825       744.9457       768.4256
#> [42,]       902.7139       946.1529       892.9410       902.8751
#> [43,]       621.0000       621.0000       621.0000       621.0000
#> [44,]      1005.8134       944.1505       976.8940      1006.8101
#> [45,]      1050.5889      1208.7150      1112.8573      1026.0176
#> [46,]      1110.6308      1064.0595      1035.6353      1071.5970
#> [47,]       875.0000       875.0000       875.0000       875.0000
#> [48,]       526.0000       526.0000       526.0000       526.0000
#> [49,]       515.9279       448.3636       500.2307       544.1037
#> [50,]      1507.8152      1531.6430      1488.3937      1567.0617
#> [51,]      1363.0000      1363.0000      1363.0000      1363.0000
#> [52,]      1099.0646      1122.1181      1140.1346      1165.6774
#> [53,]       655.6631       626.4257       627.7772       696.3059
#> [54,]       405.0000       405.0000       405.0000       405.0000
#> [55,]       658.0000       658.0000       658.0000       658.0000
#> [56,]       996.0000       996.0000       996.0000       996.0000
#> [57,]      1355.6644      1341.9050      1323.1260      1387.6118
#> [58,]      1548.0000      1548.0000      1548.0000      1548.0000
#> [59,]       772.0000       772.0000       772.0000       772.0000
#> [60,]       590.5856       647.4067       658.4874       659.8481
#>       Imputation0017 Imputation0018 Imputation0019
#>  [1,]       738.7338       726.3210       728.7715
#>  [2,]      1067.0000      1067.0000      1067.0000
#>  [3,]      1989.3143      2026.1021      2067.8230
#>  [4,]      1291.1363      1338.1904      1259.3492
#>  [5,]       946.6657       842.8117       940.2037
#>  [6,]       697.0000       697.0000       697.0000
#>  [7,]       783.8549       834.4632       781.6655
#>  [8,]      1035.9830      1103.8968      1080.5993
#>  [9,]      1821.0000      1821.0000      1821.0000
#> [10,]      1491.4572      1552.5197      1488.5066
#> [11,]      1090.7591      1084.1834       971.9762
#> [12,]       654.9439       630.8406       586.6854
#> [13,]       933.2566       893.8548       987.6105
#> [14,]      1400.0275      1435.0666      1391.1023
#> [15,]      1133.4538      1174.2747      1117.8469
#> [16,]       965.0000       965.0000       965.0000
#> [17,]       444.4995       404.0816       412.8053
#> [18,]       387.0000       387.0000       387.0000
#> [19,]       507.9464       511.6933       528.5907
#> [20,]      1137.0000      1137.0000      1137.0000
#> [21,]      1083.3817      1128.9490      1106.0187
#> [22,]      1239.8899      1282.9062      1215.2118
#> [23,]      1210.6404      1176.8412      1277.2041
#> [24,]       604.5069       520.2017       488.1031
#> [25,]      1112.8337      1060.6292      1079.6510
#> [26,]      1746.0000      1746.0000      1746.0000
#> [27,]      1334.0000      1334.0000      1334.0000
#> [28,]      1166.7804      1127.6374      1160.3206
#> [29,]       584.0000       584.0000       584.0000
#> [30,]       649.2854       631.5079       695.9504
#> [31,]       785.0000       785.0000       785.0000
#> [32,]      1291.0000      1291.0000      1291.0000
#> [33,]      1912.0000      1912.0000      1912.0000
#> [34,]      1832.0000      1832.0000      1832.0000
#> [35,]      1441.0126      1484.3318      1492.2947
#> [36,]       681.8295       610.5560       699.5633
#> [37,]       832.5401       822.9822       798.8032
#> [38,]      1387.8170      1382.2515      1417.8331
#> [39,]      1125.0000      1125.0000      1125.0000
#> [40,]      1200.0000      1200.0000      1200.0000
#> [41,]       801.3724       781.7383       843.9341
#> [42,]       880.4329       824.3995       899.4606
#> [43,]       621.0000       621.0000       621.0000
#> [44,]      1046.8741      1006.1536      1020.6185
#> [45,]      1131.6942      1126.0651      1142.1394
#> [46,]      1062.8018      1105.7430      1021.0993
#> [47,]       875.0000       875.0000       875.0000
#> [48,]       526.0000       526.0000       526.0000
#> [49,]       488.4032       469.5937       520.7041
#> [50,]      1468.4070      1528.1518      1512.3829
#> [51,]      1363.0000      1363.0000      1363.0000
#> [52,]      1150.7099      1105.8309      1120.2139
#> [53,]       632.7313       660.4897       649.3671
#> [54,]       405.0000       405.0000       405.0000
#> [55,]       658.0000       658.0000       658.0000
#> [56,]       996.0000       996.0000       996.0000
#> [57,]      1349.1546      1351.2851      1394.5557
#> [58,]      1548.0000      1548.0000      1548.0000
#> [59,]       772.0000       772.0000       772.0000
#> [60,]       703.0794       562.7616       632.2358
#>