Model an imputed dataset

Model an imputed dataset

model_impute(object, model.fun, rhs, model.args, extractor, extractor.args,
  filter, mutate)

# S4 method for ANY
model_impute(object, model.fun, rhs, model.args, extractor,
  extractor.args, filter, mutate)

# S4 method for aggregatedImputed
model_impute(object, model.fun, rhs, model.args,
  extractor, extractor.args, filter, mutate)

Arguments

object
the imputed dataset

model.fun
the function to apply on each imputation set

rhs
the right hand side of the model

model.args
an optional list of arguments to pass to the model function.

extractor
a function which return a matrix or data.frame. The first column should contain the estimate, the second the standard error of the estimate

extractor.args
an optional list of arguments to pass to the extractor function.

filter
an optional argument to filter the raw dataset before aggregation. Will be passed to the .dots argument of filter_

mutate
an optional argument to alter the aggregated dataset. Will be passed to the .dots argument of mutate_. This is mainly useful for simple convertions, e.g. factors to numbers and viceversa

Examples

dataset <- generateData(n.year = 10, n.site = 50, n.run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggr <- aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
extractor <- function(model){
  summary(model)$coefficients[, c("Estimate", "Std. Error")]
}
model_impute(
  object = aggr,
  model.fun = lm,
  rhs = "0 + factor(Year)",
  extractor = extractor
)

#> # A tibble: 10 x 5
#>    Parameter      Estimate    SE   LCL   UCL
#>    <fct>             <dbl> <dbl> <dbl> <dbl>
#>  1 factor(Year)1     1092.  166.  767. 1417.
#>  2 factor(Year)2      955.  166.  629. 1280.
#>  3 factor(Year)3     1252.  166.  927. 1578.
#>  4 factor(Year)4     1154.  166.  829. 1480.
#>  5 factor(Year)5     1313.  166.  987. 1638.
#>  6 factor(Year)6     1116.  166.  791. 1442.
#>  7 factor(Year)7     1073.  166.  748. 1399.
#>  8 factor(Year)8     1462.  166. 1136. 1788.
#>  9 factor(Year)9     1549.  166. 1223. 1875.
#> 10 factor(Year)10    1208.  166.  883. 1534.